+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/Versions.kt
@@ -60,6 +60,8 @@ internal object SentryModules {
     DefaultModuleIdentifier.newId("io.sentry", "sentry-android-navigation")
   internal val SENTRY_ANDROID_TIMBER =
     DefaultModuleIdentifier.newId("io.sentry", "sentry-android-timber")
+  internal val SENTRY_ANDROID_REPLAY =
+    DefaultModuleIdentifier.newId("io.sentry", "sentry-android-replay")
   internal val SENTRY_ANDROID_DISTRIBUTION =
     DefaultModuleIdentifier.newId("io.sentry", "sentry-android-distribution")
   internal val SENTRY_OKHTTP = DefaultModuleIdentifier.newId("io.sentry", "sentry-okhttp")